Feature/8 implement machine creation persistence (#10)
* Start working on domain model for machine definitions.

Work on #8

* Code clean-up

Work on #8

* Fix broken tests.

Work on #8

* Redesign of API.

Change idea of the API structure. Separate creation of machine from creation
of machine version.

Work on #8

* Introduce endpoint for managing nodes.

Nodes will be added through API and than interconnected through separate
endpoint for transitions.

Work on #8

* Add definition for transitions adding.

Work on #8

* Add endpoint for confirming machine.

Current idea is to block modifications of confirmed/accepted version. But
the idea is to be considered at later point.

Work on #8

* Add next unit tests.

Work on #8

* Adjust debugger display attributes.

Adjust attributes to reflect recent changes.

Work on #8

* Add machine definition class.

Work on #8

* Implement nodes adding.

Work on #8

* Organize resources.

Set naming convention for resources.

Work on #8

* Add tests for transition adding.

Work on #8

* Add status representation.

Representation of machine structure status.

Work on #8

* Status code cannot be white-space only.

Work on #8

* Add initial work on aggregates repository.

Work on #8

* Start configuration of EventStoreDB connection.

Work on #8

* Store machine details in database.

Work on #8

* Start using test containers for integration tests.

Work on #8

* Implement version creation on EventStoreDB

Switch to identification of streams by public identifier.
Implement adding a new version.

Work on #8

* Enable swagger

Add support for swagger to be able to do tests on running API.

Work on #8

* Handle MachineName serialization

Since machine name is domain object without public constructor,
serialization has to be handled in converter.

Work on #8

* Remove not needed try-catch block.

Work on #8

* Store transitions for machine.

Work on #8

* Implement machine confirmation.

Work on #8

* Implement machine definition retrieval

Work on #8